Update patch set 1
Patch Set 1:

I’m not sure if I understand correctly how this is intended to work, but as far as I can tell unsetting a reference to an array element doesn’t remove the element from the array.

 $ php r '$arr = [ 1, 2 ]; foreach ( $arr as &$el ) { if ( $el % 2 === 0 ) { var_dump( $el ); unset( $el ); } }; var_dump( $arr );'
 Command line code:1:
 int(2)
 Command line code:1:
 array(2) {
   [0] =>
   int(1)
   [1] =>
   int(2)
 }

So I think this will actually attempt to fetch everything from the injected lookup, even entity IDs that we already have cached data for.

Patch-set: 1
Label: Code-Review=0