Update patch set 62
Patch Set 62:

(11 comments)

Previously we (namely Lydia and myself) said we will allow this code to be merged, and continue working on it later. Unfortunately I do see code smells in the proposed code that are so significant and so huge that it will be much more trouble fixing them later than just not allowing them in the codebase.

Yes, this is basically a -1: I do see issues in this code that are so significant that I'm not willing to accept them as they are right now. I do have a very clear idea how all of these issues can be fixed, and I hope my comments explain enough so others can understand what I mean. If not, please ask.

I'm also willing to work on follow up patches, but I would like to have a clear way forward first: Who's responsibility is it to care about this code? Am I supposed to spend time on this, and if so, how much, and on what cost unit? Who's responsibility is it to fix bugs or cross-compatibility issues (e.g. with core or the Cirrus extension) in this code when they arrise?

Also I strongly suggest (again) to split this patch into smaller ones. For example, the first patch can move the existing code from RepoHooks::onShowSearchHitTitle to ShowSearchHitHandler::onShowSearchHitTitle.

I'm not setting a -1 because of what we previously talked about, even if this review deserves one.

Patch-set: 62