Update patch set 1
Patch Set 1:

This is a very crazy class that does very crazy things. Or in other words, I can not tell if what this patch does is correct, what specification or definition of "special" is used here, how this is used, what is fed in, and so on.

In other words: Can you please add a test?

Patch-set: 1
Label: Verified=0