Update patch set 7
Patch Set 7: Code-Review+1

(2 comments)

The code looks pretty good to me. Not +2 yet until I test it locally for more than 30 seconds. Should happen tomorrow morning Europe time.

DescriptionTest is pretty nasty in terms of the mock construction, but looking at the test cases it was clear to me what is tested etc. So the implementation of the test could still be improved in the future if we ever have a better idea.

Patch-set: 7
Label: Code-Review=+1