Handle redirects correctly in loadLatestRevisionIds
If a page is a redirect, the latest revision ID of that page is not the
latest revision ID of the corresponding entity. Return false in that
case instead of the latest revision ID of the redirect page. (This was
to some degree implied by the old documentation of the method, but the
documentation is also updated to clarify this point.)

This makes WikiPageEntityMetaDataAccessor::loadLatestRevisionIds more
consistent with WikiPageEntityRevisionLookup::getLatestRevisionId.

Bug: T188381
Change-Id: I79f98f91e8ff908e020ccc3b83c0cbba329ffd89