Update patch set 67
Patch Set 67:

> I'm also willing to work on follow up patches, but I would like to have a clear way forward first: Who's responsibility is it to care about this code? 

Our team will be responsible for the Elastic parts. Wikidata team will be taking care of the rest as usual :) If some things are in gray area, I expect we will discuss them and figure out who does what.

> Am I supposed to spend time on this, and if so, how much, and on what cost unit? 

I would appreciate the time on review (yourself or anybody else in Wikidata team, that's not for me to decide). No idea about cost units. 

> Who's responsibility is it to fix bugs or cross-compatibility issues (e.g. with core or the Cirrus extension) in this code when they arise?

If there are problems with Cirrus/Elastic part, we will be fixing them (me or somebody else from the team if I can not for some reason). 

> Also I strongly suggest (again) to split this patch into smaller ones. For example, the first patch can move the existing code from RepoHooks::onShowSearchHitTitle to ShowSearchHitHandler::onShowSearchHitTitle.

I moved the Link/hooks part into a separate patch, but I don't really want to split it any more as having a string of patches to maintain in sync for several months is very tedious, and this patch does largely one thing, which requires all its components - backend and display - to work. I know it's big, this is because it's a big change in search. Doing it is small pieces would create extra work for a project that is taking quite a long time anyway.

Patch-set: 67
Label: Verified=0