Update patch set 6
Patch Set 6:

(4 comments)

Thanks for working on this. The approach of extracting code to be able to reuse and test it separately is very much welcome! I did a full review and found a few issues that are worth improving, in my personal opinion. However, I'm not going to set a -1 for the moment as the responsibilities for this code are not as clear as I wish they would be here in the Berlin office.

Patch-set: 6
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0