Update patch set 62
Patch Set 62:

(2 comments)

As argued before, splitting of unrelated code formatting or the extraction of a hook handler function to separate patches would certainly help. All this would be merged within hours, as there is nothing really controversial about consistent code formatting or moving existing code without changing it's functionality. The impact on this patch here will not be significant, sure. It will maybe shrink by a few dozen lines. But that might already be worth it.

Patch-set: 62