Fix UsageDeduplicator accidentially being called way to often
In T189409 Anomie found a significant performance issue in the
addStatementUsage function in Lua. I could not believe this, started
digging into the code, and found this actual performance issue that
was accidentially introduced in Ic101f58 (T178079).

The deduplication is triggered *every* time a usage is added. This
does not have any effect on the result, but creates more and more
heavy CPU load, the more usages are tracked.

Note this is currently not testable because the UsageAccumulator is
not injected. Do you think a test is critical? Tell me, and I will
add a patch introducing a test on top of this one.

Bug: T189409
Change-Id: Ie1a29e2b817b8cd589f505a51f2183ee3771ca84