Micro-optimization of EntityUsage::getIdentityString
This is a quite minor performance-optimization I found while looking into
T189409, profiling and looking for possible bottlenecks in the code. The
argument is quite simple: getIdentityString is called at least two times on
each EntityUsage object:
* One time when collecting usages in ParserOutputUsageAccumulator::addUsage.
* A second time when deduplicating all usages in UsageDeduplicator.
This alone makes it worth caching this string.

I also looked into caching EntityUsage::getAspectKey, but this doesn't seem
to be worth it.

Bug: T189409
Change-Id: Ic942b698c26f980944af46e573bad0ee0004fe87