Test UsageDeduplicator not beeing called more than once
I tried to fix the TODO and always inject the UsageDeduplicator, but
realized it is constructed in 6 (!) different places in production code
(plus 20 in test). Changing them all to refer to a single top-level
factory would probably be a 200+ lines patch. I don't have the time at
the moment.

This patch here is at least a step forward. The next step could be to
make the constructor parameter a required one.

Bug: T189409
Change-Id: I694a759d4ddf35e576bb4f597cd8b2739fda2151