[WIP] Module registration: Keep global var handling central
For the most part, it was already working this way; each resource file
returns an array to be merged by the includer into a larger array, until
eventually at the end it is add it all to the wgResourceModules and
wgMessagesDirs global variables from the main entry point file of
one of the four virtual MediaWiki extensions:

- Wikibase Client (client/WikibaseClient.php)
- Wikibase Repo (repo/Wikibase.php)
- Wikibase Lib (lib/WikibaseLib.php)
- Wikibase View (view/WikibaseView.php)

However, there were a few spurious cases in view/ that were interacting
with these globals directly. This commit fixes those to more closely
resemble the other four virtual extensions.

* Convert to array-returning files.
* Moving the final merge to the entry point (view/WikibaseView.php).
* Remove view/init.mw.php (essentially renamed to view/WikibaseView.php).

Bug: T189538
Change-Id: I2cd1db24b3e2d174b4262ddff69445e89bc064e5