Update patch set 1
Patch Set 1:

To be honest I don't see much of a change in this patch here. The commit message talks about "interacting with globals", but the two "resources" files this patch deletes did not interacted with globals, and the code that did just moved to an other file. What is the benefit?

Patch-set: 1