Update patch set 2
Patch Set 2:

I updated this patch to include a second module, and fixed all issues so it succeeds now. This is still WIP because a lot more simplifications could be done. But I do not want to do this on my own. Leszek, I believe this should be your decision (a.k.a. the decision of your team). You introduced this and I remember an argument about consistency. This patch introduces a minor inconsistency in so far that two modules do not have a Wikibase/view/lib/resources/…/ subdirectory any more, but are defined in Wikibase/view/lib/resources/….php directly. Personally I'm fine with either. Please feel free to do whatever you like with this patch.

Patch-set: 2
Label: Verified=0