Fix UsageTrackingIntegrationTest if P[123] exists
In this test the actual wiki database is used to find
out whether a certain property exists. Due that
StatementTransclusionInteractor::render registers a statement
usage (C) along with the other usage, if the referenced entity
exists on the production system.

The underlying problem is not trivial to fix, so I went with this
simple workaround instead.

Change-Id: I55de8385c61536cf40ed7223af4192795eef4d38