Update patch set 74
Patch Set 74:

(2 comments)

The static Utility class is not something I find the most beautiful, but let's maybe wait until the related work is done to think of way how to re-do it.
Also the tests are not perfect to my taste, but as I don't know the intricacies of the code that good, I have no particular suggestions.

Overall, seems good enough to merge.
I cannot judge on Cirrus/Elastic-specific stuff, but I assume this has already been thoroughly reviewed and tested.

Patch-set: 74
Label: Code-Review=0