Update patch set 3
Patch Set 3: Code-Review+1

dir="auto" just tells the user agent to use the lang="…" attribute to determine the directionality. The user agent basically applies the same algorithm as the Language class does (based on a database of language codes that are known to be right-to-left). We don't need to hard-code the directionality in all cases – especially not if it's potentially as expensive as it is in this case. Clients are typically able to do this on their own.

TL;DR: I'm all for getting rid of as many Language::factory() calls as possible.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1