Update patch set 5
Patch Set 5:

> I am not sure it's a good idea to generate random data in dumps.

Yeah, that's not so nice… agreed.

We could make it depend on all relevant parameters in question, like: "$firstPageId-$lastPageId-$shardNumber-$shardCount-$whateverElse" … then hash it and take the first n chars. I would like to avoid having to come up with a --part-id in the bash script.

Patch-set: 5
Label: Verified=0