Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > I am not sure it's a good idea to generate random data in dumps.
> 
> Yeah, that's not so nice… agreed.
> 
> We could make it depend on all relevant parameters in question, like: "$firstPageId-$lastPageId-$shardNumber-$shardCount-$whateverElse" … then hash it and take the first n chars. I would like to avoid having to come up with a --part-id in the bash script.

Why not? Having external parameter is much more flexible and makes the PHP code not depend on the particulars of which dumping scheme we are using. In a month, we may decide to use different partitioning scheme. The script that implements the scheme is exactly the right place to contain the code that generates the IDs prefixes, and the generating script IMO should be ignorant of such details - it should just get the parameters of what to generate and generate a well-defined repeatable result. I think using --part-id from the script is a most flexible and future-proof solution.

Patch-set: 5
Reviewer: Smalyshev <2096@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0