Update patch set 5
Patch Set 5: Code-Review+1

Personally I'm unable to merge this patch, because it is basically mine now since I updated it so heavily in PS5. Raz, do you think your team can pick this up? This is relevant for production because of the performance it drains for every single call on a wiki with Wikibase installed, even when no Wikibase code is going to be executed.

Patch-set: 5
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1