Update patch set 4
Patch Set 4:

I'm not so sure what this patch proposes would be the best possible approach. The current way the ReadModeViewFactory works is more a quick hack I introduced one day. This class overrides the ControllerViewFactory and disables (!) some editing features. This is not a nice architecture.

Sure, extensions do not need to follow this architecture.

But one thing extensions can quite easily do is to implement their own isEditable() check, and depending on that return a ControllerViewFactory with editing features disabled. The major benefit of this is that the entityViewInit script touched here can live with a single entry point for extensions, and does not need two.

Patch-set: 4