Update patch set 1
Patch Set 1:

We could go even further and remove the Hook in in RepoHooks by declaring the list of statement boosts in the RescoreFunctions directly. I think it was not done from the beginning because it was impossible to customize Cirrus profiles.
I'll send a followup patch on top of this one to discuss if this is what we want.

Patch-set: 1
Label: Verified=0