Make UnitStorageTest more flexible
The test shouldn’t assume that all UnitStorage implementations take a
single constructor argument which is a file path – it’s better to use a
separate data provider instead.

Also removes a potentially confusing assertion – we don’t really expect
storage->getConversion( 'Q1' ) to return null, we expect it to throw an
exception.

Change-Id: I156490ef113af2af7b5d53369238ddb4c78c1fd0