Update patch set 1
Patch Set 1:

> With WikibaseLexeme installed, lexeme and form IDs are accepted.
They are valid ids, so I think this is acceptable?

> Lowercase entity IDs (q1) are accepted.
Well, they can be used to retrieve entities, so they are valid:
mw.wikibase.getEntity( 'q42' ).id
Q42

> All kinds of repositories are accepted (:Q1, foo:Q1, …).
That also makes sense… although it could be argued that we want to reject unknown repos here, but I'm not sure.

Patch-set: 1