Update patch set 1
Patch Set 1:

> lexeme and form IDs are accepted.

This is intended.

> Lowercase entity IDs (q1) are accepted.

This is intended, but only true for the original Q… and P… entity types. The new entity types in the Lexeme codebase never accepted lowercase IDs.

> it could be argued that we want to reject unknown repos

I'm a bit torn on this. Personally I would argue that an ID is formally valid, even if it points to an unknown repository. It's valid, but not known. This is the same for a deleted ID on a known repository. This ID is also valid, but no entity with this ID will be found.

Patch-set: 1