[elastic] Drop rescore function hook
Since we can now move all the fine tuning in the wiki config
I think we can remove this hook and fully use Cirrus profile
management.

The drawback is that it removes the entityBoost config var which was
handy to configure boost based on statements.
But I think it's worth the price since it decrease dependency on
cirrus internals.

Bug: T182717
Depends-On: I892e1ede24144d352a54b07bb565244f2ae270c2
Change-Id: Ifbcdc21e812e32f05bc6547a3faa461933ec40f9