Update patch set 3
Patch Set 3:

> Patch Set 1:
> 
> We could go even further and remove the Hook in in RepoHooks by declaring the list of statement boosts in the RescoreFunctions directly. I think it was not done from the beginning because it was impossible to customize Cirrus profiles.
> I'll send a followup patch on top of this one to discuss if this is what we want.

We could do that but we should remember that different wikibase installs have different ids, which means we'd need separate profiles for them if we put it in the profile. The reason I've put it into the variable is that then we could have the same profile and control it via the variable. But if you see a way to achieve the same (i.e. having control over boosted IDs without blowing up the profile setup) or think duplicating the profiles is the right thing to do, let's discuss it.

Patch-set: 3
Label: Verified=0