Update patch set 7
Patch Set 7:

The by-type delegating/dispatching seems to be generally a good path. I got myself wonder though, whether it is now suggested to be done one level too high.
As this is for "use cirrus" case, shouldn't the dispatching thing be a EntitySearchElastic dispatching thing, not the generic EntitySearchHelper?
Done like here it would make sense if the intention was to e.g. have item search elastic-powered, and property (or lexeme if you think this is too drastic example) sql-powered. But as it is configured right now it is either all-sql or all-elastic. So it seems more appropriate to have the per-type dispatching on the level of the elastic-specific service. What do you think?

On the I note there is no tests for the new dispatching service. Wouldn't it make sense to add them?

Patch-set: 7