Handle normalizePageName() returning non-string values
If $siteObj is a MediaWikiSite, normalizePageName() can return false in
a number of conditions, including for pages that are simply missing on
the remote site. Since SiteLinkLookup::getItemIdForLink() expects a
string as title, and can in some implementations throw a
ParameterTypeException otherwise, make sure that we really got a string
before passing it on.

Bug: T191634
Change-Id: I140a9510b86109e74a3f67ceed2767d9f21c8dd9