Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> The by-type delegating/dispatching seems to be generally a good path. I got myself wonder though, whether it is now suggested to be done one level too high.
> As this is for "use cirrus" case, shouldn't the dispatching thing be a EntitySearchElastic dispatching thing, not the generic EntitySearchHelper?
> Done like here it would make sense if the intention was to e.g. have item search elastic-powered, and property (or lexeme if you think this is too drastic example) sql-powered. But as it is configured right now it is either all-sql or all-elastic. So it seems more appropriate to have the per-type dispatching on the level of the elastic-specific service. What do you think?
> 
> On the I note there is no tests for the new dispatching service. Wouldn't it make sense to add them?

This is not only for use cirrus case. This should be in generic case too. Eventually for T190022 we'll move out the cirrus case into an extension, and the regular case will be the non-cirrus case. But we want to keep the dispatching part in the Wikibase core. But moving out the Cirrus part only possible if dispatching is in the Wikibase core - otherwise the core will have no capability for overrides. 

> But as it is configured right now it is either all-sql or all-elastic.

It is, but doesn't have to be. Extensions can install their overrides with dispatching part. 

> On the I note there is no tests for the new dispatching service. Wouldn't it make sense to add them?

Yes, if this model is OK I'll add tests too.

Patch-set: 7