Add maintenance script to clear term_search_key+term_weight
This maintenance script can be used to clear the term_search_key and
term_weight columns of the wb_terms database if they’re not being used
anymore. The rebuildTermSqlIndex with the --rebuild-all-terms option
enabled has the same effect if the repo is configured not to use the
search keys, but this script is much simpler and more efficient for its
single purpose.

Based on rebuildTermSqlIndex and TermSqlIndexBuilder, but simplified.
For example, we don’t need an SqlEntityIdPager – we implement paging on
the term_row_id ourselves.

Bug: T191631
Change-Id: I2c539a39f6a6a206e46b7ef98e3eabafc7d758e2