Fix ModifyEntity/EditEntity parameter validation
EditEntity needs no own checking, the only custom
parameter it needs to check is data, and that is
already marked as required, so MW does this for us.

Also show proper errors if only site or title are
given, and show error if site+title or id and new
is given.

Bug: T185165
Change-Id: I65ad9e08c0ff4b745b0af2ce0aecd2439b64e552