Update patch set 2
Patch Set 2: Code-Review+1

I was a little scared when I saw this clone() call. I actually profiled this one and found it is surprisingly cheap. The Context object does not contain much that would need expensive cloning. It is basically made of references to other objects. Only these references are cloned, not the objects. (PHP doesn't do deep cloning.) So I believe this change is fine.

Also this code is never called in a loop or something, but only one time per edit.

Patch-set: 2