Update patch set 2
Patch Set 2: Code-Review+2

I had to dig deeeeep to find when this cloning was introduced. This was 2013:

https://gerrit.wikimedia.org/r/#/c/45002/3/repo/includes/EditEntity.php

But the code was fine back then. No changes have been made to the non-cloned context. This bug was accidentally introduced in a refactoring a year later:

https://gerrit.wikimedia.org/r/#/c/180479/6/repo/includes/EditEntity.php

Which means we are actually fixing bug we introduced 4 years ago. ;-)

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2