DNM WIP DRAFT Add seperate setting for enabling entity types
In combination with the following settings this makes the 'form'
entity type appear in the list of searchable entity types in
wbsearchentities.

$wgWBRepoSettings['enabledEntityTypes']['form'] = 'form';
$wgWBRepoSettings['enabledEntityTypes']['lexeme'] = 'lexeme';

Prior to this, entity types that are enabled are loaded from
entityNamespaces, however forms do not have a namespace so they
can not be loaded in this way.

After writing this I don't think this is the correct approach.

This change moves the enabeling of entity types to a new config setting,
however entitty types such as forms that don't themselves have a
namespace may still need to be mapped to one in some way.
Wikibase generally needs a concept of sub entities, and probably some
sort of entity map to represent that...

[
 'item' => 0,
 'property' => 90,
 'lexeme' => 123,
 'form' => 'lexeme'
]

This in combination with a way to get from high level entities (item /
lexeme) to their sub entities might be the solution.

Bug: T191981
Change-Id: I1d3ecd46c365d7ae4be7bf37d064c3d3de6a973d