Update patch set 1
Patch Set 1:

Problem I see with making this somehow "configurable", as in exposing/setting it using some config/settings is that it makes it look like it was totally fine to e.g. have lexemes enabled on one repo, and forms on other, or e.g. lexemes enabled and forms not. Both possibilities seem iffy to me, at least at this point of the project.

I think it would be better if it worked the way if lexemes are enabled (be it e.g. the current way by enabling it in repository definitions and specifying the namespace), forms etc are automatically also enabled.
That said I don't have a good suggestion how to do it smart, yet. I will try to come with some draft soon

Patch-set: 1
Label: Code-Review=0