EntityStore, document throwing StorageException and PermissionsError
It looks like in Wikibase currently the only thing that would be thrown
here is a StorageException.
Howere it would make sense to me to keep this consistent with the other
methods in this interface.

For example in WikibaseLexeme the deleteEntity implementation for forms
/ sub entityies (FormStore) calls saveEntity internally (because for sub
entities a delete is just a removal from an existing entity), and thus
all of the same exceptions can occour.

Change-Id: If11d7954bd32b25925283cc103f39046fed1b68c