Remove createBlacklistedItems maint script
This was used when Wikidata.org was initially setup, but it is no
longer needed and also probably shouldn't live in the Wikibase
code repo.
If we do for some reason want to keep it I can recreate it in the
Wikidata.org mediawiki extension, but I see no point in maintaining
this code.

Change-Id: I178e64c9697fe36c3ac268ffe8a0c86c11e6dbc8