Revert "Skip tests failing in CI"
This reverts commit 3060f6f716494093bd34d49a4dae74483ae5ba6c.
It looks like the pending ajax complaints were actually caused by
I7c0d02f073a2fe165a6027e3c6d1455e22c1b563 but Wikibase
thanks to the alphabet and the position of the letter W took the blame.

I think we can safely revert this, but when doing so we should keep an
eye on tests elsewhere.

Change-Id: I2ae672db2e2b68746b4f6ed3f791ccd7a3d0514e