Update patch set 3
Patch Set 3:

Reading T189744, I believe I get the idea would be to be able to specify some "filter", or "filters", in order to say e.g. "give me only lexemes that are nouns (i.e. the lexical category is Qxyz)", or "give me only items that are units (i.e. P31 says they're instances of unit)".
This seems really cool and useful! I would like to think about the idea a bit more, although I really like the concept. Will back here soon!
I am also a bit uneasy about passing those "filters" as a JSON object. It would work, but it does not seem to be the nicest structure to use for user input. Also, arguably, the existing "language", or "type" are also these kinds of filters, so others should also be treated in a similar way. That probably would take a bit of effort to make search entities API extensible in this regard. So it could be that we end up stuffing filters into "options" (or "filter") param as a JSON for now. How urgent it is to have this functionality implemented and live?

Also, on a minor note, I am not sure if "strict languages" is the "option" of the same kind as the above mentioned "filters". I might clearly lack some knowledge, but it seems to me those should be kept separate.

Patch-set: 3
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0