Update patch set 1
Patch Set 1: Code-Review+1

I more and more feel the Wikibase interface does have a bit to many borders and boxes. But this particular one is one I feel is super-helpful and pretty much necessary, in my humble opinion.

I think the color should be the same that is assigned to the "border" keyword in the image inclusion syntax. That is #eaecf0 – a different one than the one you picked, it seems. I would ask WMDE Design if in doubt.

Also it would be super-awesome if we could finally reuse the named LESS variables defined in wikimedia-ui-base.less ("oojs-ui-core.styles" as a resource loader module). We are currently discussing this in Ifdadec9. It seems it's currently not safely possible – so I bring this up to let you know, not as a request. :-)

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1