Update patch set 1
Patch Set 1: Code-Review-1

TL;DR: I think there is no way to fix the problem "right way" fast, but of all walkarounds I claim I have a better (tm) suggestion. Code coming up soon, as it is easier than me trying to describe the idea in an essay.

Aleksey is right, it is difficult to even see what this is about without test, and most importantly, the issue would be easy to accidentally be re-added if there is no tests. 

Hoo man: ChangeOpAliases actually does rather good job, but it is only handling the "atomic" operation, i.e. either only adds, or only removals. In this case there is an add, and a removal, thus two ChangeOp objects stuffed into ChangeOps which cannot do meaningful summary. To overcome this SetAliases tries to be smart an generates the summary itself. But it does it slightly wrong, and hence the bug.

Patch-set: 1
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1