Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

I'm very much willing to merge this as it is, with no further change. It's just that the text made me think a little bit to much, in my opinion, which can be avoided by rephrasing it slightly. What do you think?

Is this part of a specific story that is tracked on Phabricator?

Patch-set: 1
Label: Code-Review=+1