Update patch set 2
Patch Set 2: Code-Review+2

I agree with Aleksey that simplifying all this would make test more clear, and maybe also more reliable.

That said, as this change is blocking the change needed for the deployment we want to do today, I am going to give the bad example, and merge this as it is, and swear the oath, the I am going to re-do those tests on my next working day (to hereby declare it is going to be Fri, 11.05.2018).

I accept this exception only because of the urgent changes depending on this quick fix, and also realizing that the way MediaWiki tests are  dealing with the DB is very far from "right" (which in general is not an excuse for unstable tests in our code, to be clear on this).

Patch-set: 2
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2