Update patch set 3
Patch Set 3:

What is said in the commit message is correct. However, I don't think this is the right way to model this here. Note the number of entity types is not limited. A Wikibase repository can have any number of entity types (e.g. "query", "lexeme", "mediainfo" and so on). With the approach currently presented in this patch you would need to add every entity type to every other entity type, every time a new entity type is added to the system. Who is going to maintain this, and how?

Isn't there a way to say that all subclasses of an entity are exclusive to each other, without listing the subclasses by name?

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0