Update patch set 2
Patch Set 2:

So apparently the behaviour which I claim was invalid here was never really surfacing on items and properties, but we hit it the request like in https://gerrit.wikimedia.org/r/#/c/433172/3/tests/phpunit/mediawiki/Api/LexemeEditEntityTest.php@605.

The point in that case: validate method of our change op class has a check if the form with ID given exists on the lexeme, and does not check otherwise. Then in the apply we don't repeat the check when getting form data with the said ID from the lexeme (as we expect validate to have taken care of it already).

In case clear is provided to wbeditentity, the code on ModifyEntity runs apply on the cleared entity (i.e. on the lexeme with no forms), but before it calls validate on different entity, loaded from DB (i.e. not cleared one). In our case the loaded lexeme has a form, so validate says "go ahead", and then apply blows up.

I am going to try adding a test showing off this behaviour in the core Wikibase. Hopefully this makes this rather corner case situation more clear.

Patch-set: 2