Update patch set 4
Patch Set 4: Code-Review-1

What I recently understood is that this error is not actually an error.
It is not like something went wrong and we need to try again, it means that we decided to wait X seconds and it wasn't enough. 

Regarding that, I would say that retry is a wrong thing to do.
If we say that X seconds is not enough and we should wait longer - lets increase `innodb_lock_wait_timeout`.
If we think that X seconds is enough, then we need to investigate why is this happening and what is the cause of that. May be, lets not acquire locks for such a long time.

This solution, for example, might lead to an issue when we will be waiting more than webserver's timeout and in the end user will see `504 Gateway Timeout` but request will be handled, so user will repeat her action and there will be duplicate.

Blocking the patch for the reasons above. If majority of WD team decides that this approach is fine, feel free to remove my vote.

BTW, does anyone know the value of `innodb_lock_wait_timeout` on the production server? I think we need to know it to make a right decision.

Patch-set: 4
Reviewer: Aleksey Bekh-Ivanov (WMDE) <4089@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1