Update patch set 4
Patch Set 4:

(1 comment)

If you search for "Wvzu8gpAAEkAAJkU5NAAAABQ" in logstash you will see interesting entries. Perhaps the counter increment could be moved towards the end of the transaction. The actually wb_changes INSERT seems to be near the end anyway (at a glance).

This seems to use savepoints currently. If the goal is to catch lock timeouts, then that can be done without checking SQL error codes (which is strongly discouraged). On exception, if cancelAtomic() succeeds, then a retry of the atomic section can be attempted. That said, I'd prefer not to use savepoints at all here if the SQL writes can just be moved around. Another option, failing that, is using DBO_TRX_AUTOCOMMIT for the connection to do these table increments (e.g. out of the main transaction round and in auto-commit mode). It would increase the case where an extra DB connection is used though (these are at least reused among callers though).

----

0	0.000284	query-m: SAVEPOINT `wikimedia_rdbms_atomic1` [TRX#a3f002]
1	0.000388	query-m: UPDATE `wb_id_counters` SET id_value = 'X' [TRX#a3f002]
2	0.000254	query-m: RELEASE SAVEPOINT `wikimedia_rdbms_atomic1` [TRX#a3f002]
3	4.831752	...delay...
4	0.000416	query-m: INSERT IGNORE INTO `page` (page_namespace,page_title,page_restrictions,page_is_redirect,page_is_new,page_random,page_touched,page_latest,page_len) VALUES ('X') [TRX#a3f002]
5	0.001637	query-m: INSERT INTO `blobs_cluster2N` (blob_text) VALUES ('X')
6	0.000357	query-m: INSERT INTO `text` (old_id,old_text,old_flags) VALUES (NULL,'X') [TRX#a3f002]
7	0.000456	query-m: INSERT INTO `revision` (rev_page,rev_parent_id,rev_text_id,rev_minor_edit,rev_timestamp,rev_deleted,rev_len,rev_sha1,rev_comment,rev_user,rev_user_text,rev_content_model,rev_content_format) VALUES ('X',NULL,NULL) [TRX#a3f002]
8	0.000329	query-m: INSERT INTO `revision_comment_temp` (revcomment_rev,revcomment_comment_id) VALUES ('X') [TRX#a3f002]
9	0.464057	...delay...
10	0.000452	query-m: UPDATE `page` SET page_latest = 'X' [TRX#a3f002]
11	0.000336	query-m: DELETE FROM `redirect` WHERE rd_from = 'X' [TRX#a3f002]
12	0.000478	query-m: INSERT INTO `wb_changes` (change_type,change_time,change_object_id,change_revision_id,change_user_id,change_info) VALUES ('X') [TRX#a3f002]
13	0.000354	query-m: UPDATE `user` SET user_editcount=user_editcount+N WHERE user_id = 'X' AND (user_editcount IS NOT NULL) [TRX#a3f002]

Patch-set: 4
Label: Code-Review=0