Update patch set 3
Abandoned

I guess it doesn't need to be my problem any more if people want to have methods in this codebase that are literally called "dummy" and hide any intend.

For reference, the code in question is now in \Wikibase\DataAccess\Tests\MultiRepositoryServicesTest.

Patch-set: 3
Status: abandoned
Tag: autogenerated:gerrit:abandon