Use Kartographer mapframes for displaying maps
Note: This needs https://gerrit.wikimedia.org/r/433890 or
an equivalent solution to work properly when editing.

Tested with and without Kartographer, on desktop and mobile.

Known problem: If there are multiple maps on a page,
and one is being edited, all maps (but the one to be
edited) will turn white. This can be addressed in a
follow up (change Kartographer so that we no longer
need to fire the "wikipage.content" hook).

Bug: T184933
Depends-On: I4800e388a25d049563775e1ea9dc53a4e94c8a26
Change-Id: I3da2920d4603495c03c5c9683671857a61021f3d