Update patch set 3
Patch Set 3:

@Thiemo I consider this change also to help resolve T192623, which is about reducing the number of file bundles (modules) being registered by Wikibase.

By being in centralised larger files, it will be easy to consolidate things as needed. The current pattern is problematic because it hides the cost of registering modules by giving the appearance that everything is isolated and registered as little units.

Ultimately, it is important for Wikibase to embrace the fact that the module loader is not a class loader. We should treat modules similar to how one in other projects might treat Webpack entrypoints for bundle generation. Each module is a deliverable bundle that is registered to the client up front and separately loadable. Unless a module brings in a significant cost, it is usually not worth registering as its own module, in favour of having slightly larger modules, and therefore better caching across views, a much simpler bundle graph, and a much smaller footprint on all page views.

Right now, no matter what URL is viewed on a MediaWiki instance, there is 14 KB being downloaded on every view everywhere *just* to register all the Wikibase modules. No actual code in that payload. This is a problem :)

Patch-set: 3
Label: Verified=0