Update patch set 5
Patch Set 5:

(1 comment)

I am not very fond of the "architecture" side of things here, as also mentioned in the inline comment. This change would be fine though, if there was a more or less clear idea how to change it _soon_ (like in a week or so), and it was done like this only to get the Special:ItemDisambiguation working again. If this is the plan, please proceed!

But in general I cannot help the feeling that this looks like the part of shooting at random strategy, which I think we should no longer continue with the whole of wb_terms stuff. It might help in some area, but also at very least adds loads of confusion at the same time (see e.g. the not clear behaviour for non-items).

Regarding the actual change of the behaviour of the special page: that should be Product decision, but again, I think introducing the undefined (or "not useful", if one prefers) behaviour would only make sense if this was clearly made as a temporary "hack" to get the special page back up.

Patch-set: 5
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0